backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for a personal expense tracker."
authors = [
    { name = "Hung Dao", email = "hungdd2604@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "python-dotenv",
    "alembic",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "httpx>=0.28.1",
    "google-auth",
    "google-auth-oauthlib",
    "python-multipart==0.0.20",
]

[tool.setuptools]
py-modules = ["app", "models", "routes"]
